mining: Move txPriorityQueue to a separate file.
This moves txPriorityQueue and the related logic to a separate file,
mining/txpriorityqueue.go.  Additionally, it moves the txPriorityQueue
tests to mining/txpriorityqueue_tests.go.

This is part of an effort to split mining into several files to make the
package easier to navigate, and in preparation of adding additional test
coverage.
